--- src/main/webapp/WEB-INF/views/coffeeclass/class_main.css
@@ -1,175 +1,0 @@
-@charset "UTF-8";
-
-@font-face {
-    font-family: 'NEXON Lv1 Gothic OTF';
-    src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_20-04@2.1/NEXON Lv1 Gothic OTF.woff') format('woff');
-    font-weight: normal;
-    font-style: normal;
-    }   
-
-
-    *{
-    font-family: 'NEXON Lv1 Gothic OTF';
-    margin : 0;
-    padding : 0;
-    box-sizing: border-box;
-}
-
-/* .navbar{
-    display : flex;
-    align-items: center;
-    padding: 15px;
-}
-
-nav{
-    flex: 1;
-    text-align: right;
-}
-
-nav ul{
-    display: inline-block;
-    list-style-type: none;
-}
-
-nav ul li{
-    display : inline-block;
-    margin-right: 10px;
-}
-
-nav ul :last-child{
-    margin-right: 0;
-}
-
-a{
-    text-decoration: none;
-    color : #555;
-}
-
-a:hover{
-    color:crimson;
-}
-
-p{
-    color : #555;
-} */
-
-
-
-.container{
-    margin: auto;
-    max-width: 1300px;
-}
-
-.row{
-    display : flex;
-    align-items: center;
-    flex-wrap: wrap;
-    
-}
-
-.col-2{
-    flex-basis: 50%;
-    min-width: 300px;
-}
-
-.col-2 img{
-    max-width: 100%;
-    padding: 50px 0;
-}
-
-.classes{
-    margin: auto;
-    padding-left : 30px;
-    padding-right: 30px;
-}
-
-.classes h2{
-    margin: 80px;
-    text-align: center;
-    position: relative;
-    line-height: 60px;
-}
-
-.classes h2::after{
-    content : '';
-    background-color: rgb(194, 182, 160);
-    width: 80px;
-    height: 3px;
-    border-radius: 2px;
-    position: absolute;
-    bottom : 0;
-    left : 50%;
-    transform: translateX(-50%);    
-}
-
-.col-5{
-    flex-basis : 20px;
-    padding: 10px;
-    min-width: 20%;
-    min-height: 500px;
-    margin-bottom : 50px;
-    transition: transform 0.5s;
-}
-
-.col-5 img{
-    width : 100%;
-}
-
-.rating .fa{
-    color : rgb(170, 113, 29);
-}
-
-.col-5:hover{
-    transform: translateY(-15px);
-}
-
-.input{
-    width : 380px;
-    height : 38px;
-    border-radius: 30px;
-}
-
-:root{
-    --clr-primary: #ee6352;
-    --clr-body: #333;
-    --clr-bg:#ddd;
-}
-
-.search_box{
-    padding : 60px;
-    padding-top: 100px;
-    top: 600%;
-    left : 20%;
-    height: 30px;
-    text-align: center;
-}
-
-#submitbtn{
-    cursor: pointer;
-    border: 0;
-    background : 0;
-    border-radius: 50%;
-}
-
-.filter{
-    text-align: center;
-    
-}
-
-.filter select{
-    border: 0;
-   margin : 15px;
-}
-
-
-#filterbtn {
-    background-color: transparent;
-    line-height: 25px;
-    width : 100px;
-    border:0px;
-}
-#filterbtn i:hover {
-position :relative;
-left : 20%;
-transition: all 0.25s linear 0.2s;
-}
