+++ src/main/webapp/resources/css/coffeeclass/editclass.css
@@ -1,0 +1,241 @@
+@charset "UTF-8";
+
+        /* 이랜드 나이스체 */
+        @font-face {
+            font-family: 'ELAND_Nice_M';
+            src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts-20-12@1.0/ELAND_Nice_M.woff') format('woff');
+            font-weight: normal;
+            font-style: normal;
+        }   
+
+        *{
+            font-family: 'ELAND_Nice_M';
+        }
+
+       
+        .banner{
+            width : 100%;
+            height: 80vh;
+            background-image: url('../../../resources/img/coffeeclass/createbanner.png');
+            background-repeat: no-repeat;
+            background-size: cover;
+            text-align: center;
+            vertical-align: middle;
+        }
+
+        .intro{
+            color : #e3a259;
+            text-align: center;
+            padding: 20px;
+            font-family: 'ELAND_Nice_M';
+        }
+
+        #intro-sub{
+            color: black;
+        }
+        
+        .classabout{
+            text-align: center;
+        }
+
+        #createform{
+            text-align: center;
+			border-radius: 30vw; 
+			background-color : #f9f2fb;"
+        }
+
+        .classdes{
+            width : 680px;
+        }
+
+        .answer{
+            width : 300px;
+            height: 30px;
+        }
+
+
+        .btnarea{
+            margin-top: 5%;
+            margin-bottom: 5%;
+            text-align: center;
+        }
+
+        .classdate{
+            width : 150px;
+            text-align: left;
+        }
+
+        #tbColor{
+            background-color : #fddfa6;
+            padding-top: 5%;
+            margin-top: 2%;
+            padding-bottom: 3%;
+        }
+
+        #clTable{
+            margin-left: auto;
+            margin-right: auto;
+            /* margin-top: 10%; */
+        }
+
+        #clTable td{
+            padding-top: 1%;
+            /* padding-left: 7%;
+            padding-right: 7%; */
+        }
+
+        #clTable input[type=text]{
+            border-radius: 1vw;
+            border-color: #d8d4d070;
+        }
+
+        #linkedcafe{
+            padding-top: 1vh;
+             margin-left: 4vh; 
+             width :200px; 
+             float: left; 
+             border: transparent; 
+             background-color: transparent;
+        }
+
+        /* 버튼 색깔 */
+        #register{
+            color: white;
+            background-color: #e3a259;
+        }
+        #register:hover{
+            color: black;
+            background-color: #f1d1ac;
+        }
+        #cancel{
+            color: white;
+            background-color: #f0c21b;
+        }
+        #cancel:hover{
+            color: black;
+            background-color: #f8e18d;
+        }
+
+        #postcode-btn{
+            color: black;
+            background-color: #f1d1ac;
+        }
+         /* 네비부분 */
+         nav{
+            background-color: transparent;
+            position: fixed;
+            z-index: 99;
+            width: 100%;
+            padding: 30px 10px;
+        }
+
+        .menuWrapper {
+            display: flex;
+            margin : 0;
+            padding : 0;
+            z-index: 10;
+            border-bottom: 1.5px solid white;
+            height: 80px;
+            align-items: flex-start;
+        }
+
+        .menuWrapper > * {
+            display: flex;
+            text-align: center;
+            box-sizing: border-box;
+        }
+
+        .menuWrapper .mainMenu {
+            width : 43vw;
+            min-width : 600px;
+            max-width : 50vw;
+        }
+
+        .mainMenu ul,
+        .loginArea .loginMenu ul {
+            padding: 0;
+            margin : 0;
+            display: flex;
+            align-items: flex-start;
+            vertical-align: middle;
+        }
+        .mainMenu ul li {
+            align-self: flex-start;
+            list-style: none;
+            text-align: center;
+            padding : 38px;
+            line-height: 1px;
+        }
+
+        .mainMenu a {
+            text-decoration: none;
+            color : white;
+            transition: all 0.5s linear;
+        }
+
+        .mainMenu a:hover {
+            /* color : #F3D798; */
+            color : #4B654A;
+            transition: all 0.5s linear;
+        }
+
+        .logoArea {
+            width : 10%;
+            min-width: 130px;
+            height: 100%;
+            text-align: center;
+            margin : 0;
+            display: flex;
+            align-items: center;
+        }
+
+        .logoArea img {
+            width : 80px;
+            height: 60px;
+            margin : auto;
+            display: block;
+        }
+
+        .loginArea {
+            width : 45vw;
+        }
+
+        .loginArea .infoArea {
+            width : 60%;
+            /* min-width: 400px; */
+        }
+
+        .loginArea .loginMenu {
+            width: 40%;
+            min-width: 200px;
+            float: right;
+            text-align: right;
+        }
+
+        .loginArea .loginMenu ul {
+            display: flex;
+            justify-content: flex-end;
+        }
+
+        .loginArea .loginMenu ul li {
+            list-style-type: none;
+            text-align: center;
+            padding : 38px 15px;
+            line-height: 1px;
+        }
+
+        .loginArea .loginMenu ul li a {
+            text-decoration: none;
+            color : white;
+            transition: all 0.5s linear;
+        }
+
+        .loginArea .loginMenu a:hover {
+            /* color : #F3D798; */
+            color : #4B654A;
+            transition: all 0.5s linear;
+        }
+
+        nav.active{
+            background: #CDC2AF;
+        }
